requirements.txt
discord.py==1.7.3
discord==1.7.3
dnspython==1.16.0
PyNaCl==1.5.0
async-timeout==3.0.1
youtube-dl==2021.12.17
pytube
ffmpeg==1.4
python-dotenv==0.19.2
youtube-search==2.1.0